refactor: Update config, chat-contacts, welcome, chat, chat-convo, chat-message, icons, chat-header, and App components
Update config.js to remove the orgId value and update it to an empty string.

Update public/appmint-chat.js to change the URL for the main chunk.

Update tailwind.config.js to remove the purge.enabled property and update the content property.

Update src/utils/index.ts to move the classNames function to the end of the file.

Update src/components/Registration/registration.tsx to remove the BuildingOffice2Icon and add the EnvelopeIcon.

Update src/chat-store.ts to import the ObjectId from the bson package and use it to generate a unique sk value for new messages.

Update update-script-hashes.js to copy the main.css file if it exists in the css directory.

Update package.json to add a prebuild script to run the tailwind:build script.

Update public/chat-demo.html to change the URLs for the favicon and logo images, and add a link to the main.css file.

Update src/App.tsx to add the BsChatLeft icon and update the styles and behavior of the chat component.